@@ -1,4 +1,4 @@
- class Game < ActiveRecord::Base
+ class Game < ApplicationRecord
   belongs_to :away_team, class_name: 'Team', foreign_key: 'away_team_id'
   belongs_to :home_team, class_name: 'Team', foreign_key: 'home_team_id'
   validates :home_team, presence: true
